+++ graph_display/lib/src/common.dart
@@ -1,0 +1,16 @@
+import 'dart:ui';
+
+import 'package:vector_math/vector_math.dart';
+
+extension OffsetExt on Offset {
+  Vector2 toVector2() => Vector2(dx, dy);
+}
+
+// TODO: See https://github.com/google/vector_math.dart/issues/189 .
+extension SetToggle<T> on Set<T> {
+  /// Toggles whether [element] is an element of this set.
+  ///
+  /// If [element] is not currently a member, it is added, and this method returns `true`.
+  /// If [element] is currently a member, it is removed, and this method returns `false`.
+  bool toggle(T element) => !remove(element) && add(element);
+}