+++ graph_layout/lib/graph_layout.dart
@@ -2,3 +2,4 @@ library graph_layout;
 
 export 'src/data_structures.dart';
 export 'src/spring.dart';
+export 'src/common.dart';