+++ graph_layout/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 * Add basic data structures: `Node`, `IntegerNode`, `Edge` (undirected), `Graph`.
 * Add basic typedefs: `AdjacencyList`, `EdgeList`, `NodeLayout`.
-* Add spring based layout algorithm `SpringSystem`, fully implemented in Dart.
+* Add spring based layout algorithm `SpringSystem`, with support for constrained (fixed) nodes.